办公室，完成控制网页

+++ soft2/STC_LIB/USART.h
@@ -40,12 +40,12 @@ typedef struct
 { 
 	u8	id;				//串口号
 
-	u8	TX_read;		//发送读指针
-	u8	TX_write;		//发送写指针
+	u16	TX_read;		//发送读指针
+	u16	TX_write;		//发送写指针
 	u8	B_TX_busy;		//忙标志
 
 	u16 	RX_Cnt;			//接收字节计数
-	u8	RX_TimeOut;		//接收超时
+	u16	RX_TimeOut;		//接收超时
 	u8	B_RX_EN;		//允许接收块
 	u8	B_RX_OK;		//允许接收块
